Athuntication

+++ Tawfeer/app/Http/Controllers/products.php
@@ -35,6 +35,7 @@ class products extends Controller
             'Products' => $jsonContent
         ]);
     }
+
     // Store new Product
     public function store(Request $request){
         //The validation
@@ -48,6 +49,7 @@ class products extends Controller
         ]);
         if($valid->fails())
             return response()->json($valid->errors()->all());
+
         //creat a new row in product table
         $product = new Product();
         $product->productName = $request->input('productName');
@@ -57,8 +59,7 @@ class products extends Controller
         $product->imgUrl = $request->input('imgUrl');
         $product->quantity = $request->input('quantity');
         $product->category = $request->input('category');
-        // Don't Forget The Owner ID from the Token in the Header
-        $product->ownerId = $request->header('ownerId');
+        $product->ownerId = auth()->user()->id;
         $product->dateOne = $request->input('dateOne');
         $product->priceOne = $request->input('priceOne');
         $product->dateTwo = $request->input('dateTwo');
@@ -67,16 +68,23 @@ class products extends Controller
         $product->priceThree = $request->input('priceThree');
         $product->save();
 
+        $seen = new Seen();
+        $seen->productId = $product->id;
+        $seen->userId = auth()->user()->id;
+        $seen->save();
+
         return response()->json(['message' => 'The Product has benn added successfully'],200);
     }
 
-
     public function show($productId){
+        // check if Wrong id
         if(!Product::where('id',$productId)->exists())
             return response()->json(['message' => 'Invalid ID'],404);
+
         // Handling the Seen
-        // Don't Forget to get the userId from the Token
-        $this->seen($productId , 1);
+        $userId = auth()->user()->id;
+        $this->seen($productId , $userId); // call seen Function
+
         // Get The Product
         $product = Product::where('id',$productId)->get([
             'productName',
@@ -103,11 +111,11 @@ class products extends Controller
     public function seen($productId , $userId){
         //Get the Product Views
         $seen = Seen::where('productId',$productId)->get();
-        $jsonContent = json_decode($seen , true);
+
         $flag = true;
-        foreach ($jsonContent as $array){
+        foreach ($seen as $seenArray){
             //check if the user had seen the product before
-            if($array['userId'] == $userId){
+            if($seenArray->userId == $userId){
                 $flag = false;
                 break;
             }
@@ -126,16 +134,30 @@ class products extends Controller
     }
 
     public function destroy($productId){
+        // check if Wrong id
+        if(!Product::where('id',$productId)->exists())
+            return response()->json(['message' => 'Invalid ID'],404);
+
         // Get the product where the id is equal to productId
         $product = Product::find($productId);
         if(!$product)
             return response()->json(['message' => 'Invalid ID']);
+
+        // check if the user has this product
+        $userId = auth()->user()->id;
+        if($product->ownerId != $userId)
+            return response()->json(['message' => 'You cant delete this product'] , 400);
+
         // Delete it
         $product->delete();
         return response()->json(['message' => 'The Product Has Been Delete successfully']);
     }
 
     public function update(Request $request,$productId){
+        // check if Wrong id
+        if(!Product::where('id',$productId)->exists())
+            return response()->json(['message' => 'Invalid ID'],404);
+
         $valid = Validator::make($request->all() , [
             'productName' => ['string'],
             'description' => ['string'],
@@ -144,7 +166,12 @@ class products extends Controller
         if($valid->fails())
             return response()->json($valid->errors()->all());
 
+        $userId = auth()->user()->id;
         $product = Product::find($productId);
+        // check ig this user has this product
+        if($userId != $product->ownerId)
+            return response()->json(['message' => 'You cant update this product'] , 400);
+
         $product->productName = !empty($request->productName) ? $request->productName : $product->productName;
         $product->description = !empty($request->description) ? $request->description : $product->description;
         $product->oldPrice = !empty($request->oldPrice) ? $request->oldPrice : $product->oldPrice;
@@ -162,7 +189,8 @@ class products extends Controller
         return response()->json(['message' => 'The Product Has Been Edit Successfully']);
     }
 
-    public function myProducts($userId){
+    public function myProducts(){
+        $userId = auth()->user()->id;
         $product = Product::where('ownerId',$userId)->get([
             'productName',
             'description',
@@ -189,6 +217,4 @@ class products extends Controller
 
 // Image URL
 // Model Binding
-// Token
-// Login
 // What to send
